requirements.txt
ImageHash==4.2.1
numpy==1.22.1
opencv-python==4.5.5.62
Pillow==9.0.0
plum-py==0.7.2
pycodestyle==2.7.0
pycparser==2.21
PyWavelets==1.2.0
scipy==1.7.3
six==1.16.0
